renamed apollo-federation/src/schema/extended_type.rs -> apollo-federation/src/schema/locations.rs
