+++ apollo-federation/src/schema/mod.rs
@@ -67,8 +67,8 @@ pub(crate) mod argument_composition_strategies;
 pub(crate) mod blueprint;
 pub(crate) mod definitions;
 pub(crate) mod directive_location;
-pub(crate) mod extended_type;
 pub(crate) mod field_set;
+pub(crate) mod locations;
 pub(crate) mod position;
 pub(crate) mod referencer;
 pub(crate) mod schema_upgrader;
